fixed the test failure caused by two variables declared as int instead of long, which caused jvm crash when offset overflowed to long

+++ core/src/main/scala/org/apache/spark/sql/execution/SHAMapAccessor.scala
@@ -391,7 +391,7 @@ case class SHAMapAccessor(@transient session: SnappySession,
            baseKeyoffset)
           }
         // insert or lookup
-        |int $valueOffsetTerm = $hashMapTerm.putBufferIfAbsent($baseKeyObject, $baseKeyoffset,
+        |long $valueOffsetTerm = $hashMapTerm.putBufferIfAbsent($baseKeyObject, $baseKeyoffset,
         |$numKeyBytesTerm, $numValueBytes + $numKeyBytesTerm, ${hashVar(0)});
         |$bbDataClass $valueDataTerm = $hashMapTerm.getValueData();
         |Object $vdBaseObjectTerm = $valueDataTerm.baseObject();
